refactor: split large MemoriesDashboardOverlay into focused modules
Break down 1203-line component into smaller, maintainable files:

- utils.ts: Utility functions (formatCount, deduplication, memory flattening)
- types.ts: TypeScript interfaces and state definitions  
- reducer.ts: State management logic with dashboardReducer
- hooks.ts: Custom business logic hooks for API calls and event handling
- SelectedSubjectBar.tsx: Reusable component for subject editing UI
- PairSearchSection.tsx: Complex component for pair search and results

Main component reduced from 1203 → 307 lines (74% reduction)
Improved maintainability, testability, and code organization
